Pushing to origin in publish_to_ghpages.sh

+++ publish_to_ghpages.sh
@@ -4,11 +4,11 @@
 
 #cd $DIR/..
 
-#if [[ $(git status -s) ]]
-#then
-#    echo "The working directory is dirty. Please commit any pending changes."
-#    exit 1;
-#fi
+if [[ $(git status -s) ]]
+then
+    echo "The working directory is dirty. Please commit any pending changes."
+    exit 1;
+fi
 
 echo "Deleting old publication"
 rm -rf public
@@ -17,7 +17,7 @@ git worktree prune
 rm -rf .git/worktrees/public/
 
 echo "Checking out gh-pages branch into public"
-git worktree add -B gh-pages public upstream/gh-pages
+git worktree add -B gh-pages public origin/gh-pages
 
 echo "Removing existing files"
 rm -rf public/*
@@ -27,3 +27,5 @@ hugo
 
 echo "Updating gh-pages branch"
 cd public && git add --all && git commit -m "Publishing to gh-pages (publish.sh)"
+
+git push origin gh-pages
